Change getApi to createApi, it exports the api
